backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-jose==3.3.0
python-multipart==0.0.6
sendgrid==6.10.0
python-dotenv==1.0.0 